Update patch set 2
Patch Set 2: No score

I personally do not allow automatic orientation changes on my phones, so I often miss bugs and suboptimal behavior like this. (I also greatly dislike the default behavior of treating activities and views as disposable entities, ready to be killed and recreated at drop off hat. It complicates the lifecycle of activities and makes reasoning about the application behavior difficult.)

I get it that you are trying to do the right thing. From user interface point of view it's rather imperative that user can hear the changes instantly, so we definitely need that, but it's also wrong to push the changes into the preferences instantly. Maybe we ought to have another solution: introduce a kind of "live demo" mode by establishing connection to HeadsetService that could update the state of the DSP without actually using the preference persisting mechanism to convey the new state.

I won't have the time to implement this right now, so if you feel like doing this then more power to you. Otherwise it'll be done by me at some point, probably. DSPManager is really my first android application and it shows...

Patch-set: 2