Update patch set 2
Patch Set 2:

In this case, I'd like you to submit a 3rd version that just unconditionally enables DSP during ringing. It should really be practically an oneliner. I don't think people are going to come up here and say that DSP must *not* be applied to their ringtone, as I have a hard time coming up with a valid use case why. However, if people do have objections then this is a good time to raise them.

It might also be possible to enable DSP during in-call, something that such a checkbox would definitely be needed for, though, but let's not do this now. I have disabled DSP during calls because I have worried about causing glitches and increased battery drain.

Patch-set: 2