Update patch set 1
Patch Set 1:

Heh, it looks like we're too late with this. The previous change already went in. Let's leave this for now -- I'll probably do an update to DSPManager at some point in the future and will review this and other changes at that time.

That being said I'd have liked the change to have "default" and "OFF_HOOK" handlers only, so that there's always a change to inCall variable on telephonymanager event.

Patch-set: 1