Update patch set 3
Patch Set 3:

Well, if the English variants are used more frequently, I would not translate them. But I do not see the technical value of the room type? We also need to let users understand it, and it would be odd if something is in English, if it is not even really technical.

Patch-set: 3