Make HeadsetService a foreground service
This service must be running all the time, so that it can respond to
AudioEffect intents and manage audio sessions. Some users complained about
DSPManager not always managing their audio sessions, and I saw that
Android sometimes kills the DSPManager process and then ends up restarting
it mere seconds later. This is a likely cause for the problems, because
DSPManager will forget the IDs of the audio sessions in a restart.

Marking the service as a foreground service should make Android not
kill the process hosting it.

This bugfix should be backported to every maintained CM release.

Change-Id: I8c9d88935e1a7403fef36bc89bce3d90e2555b1a