Update patch set 4
Patch Set 4:

It is not necessary or proper to attack the effort or work provided by contributors. This system is for a code review, not insults. 

If you spot an issue with translations (ie calculator) then let it be known. Do not dismiss the work of others who are volunteering their time and effort, it is rude and not what this community is about. 

If you feel you can truly do better, then submit a patch instead (or amend as needed). 

If this is not sufficient, then your help is not needed.

Patch-set: 4